requirements.txt
fastapi
uvicorn[standard]
huggingface_hub
langchain
langchain-anthropic
langchain-openai
pypdf
pydantic-settings
sqlalchemy
psycopg2-binary
python-dotenv
